+++ package.json
@@ -13,28 +13,32 @@
     "format:check": "prettier --check \"src/**/*.{js,jsx,ts,tsx}\" && eslint \"{src,apps,libs,test}/**/*.{js,jsx,ts,tsx}\" --max-warnings=0"
   },
   "dependencies": {
+    "@react-google-maps/api": "^2.13.1",
     "@react-navigation/material-bottom-tabs": "^6.2.4",
     "@react-navigation/native": "^6.0.13",
     "@react-navigation/native-stack": "^6.9.1",
     "expo": "46.0.16",
-    "expo-localization": "13.1.0",
+    "expo-localization": "~13.1.0",
     "expo-status-bar": "1.4.0",
     "firebase": "9.13.0",
     "i18n-js": "4.1.1",
     "react": "18.0.0",
     "react-dom": "18.0.0",
-    "react-native": "0.69.5",
-    "react-native-paper": "^4.12.5",
-    "react-native-safe-area-context": "^4.3.1",
-    "react-native-screens": "~3.18.0",
+    "react-native": "0.69.6",
+    "react-native-google-places-autocomplete": "^2.4.1",
+    "react-native-maps": "0.31.1",
+    "react-native-paper": "4.12.5",
+    "react-native-safe-area-context": "^4.4.1",
+    "react-native-screens": "^3.18.2",
     "react-native-vector-icons": "^9.2.0",
-    "react-native-web": "0.18.10",
+    "react-native-web": "0.18.9",
     "spelieve-common": "^1.1.40"
   },
   "devDependencies": {
-    "@babel/core": "7.20.2",
+    "@babel/core": "7.19.6",
+    "@types/googlemaps": "^3.43.3",
     "@types/i18n-js": "3.8.3",
-    "@types/react": "18.0.25",
+    "@types/react": "18.0.24",
     "@types/react-native": "0.70.6",
     "@types/react-native-vector-icons": "^6.4.12",
     "@typescript-eslint/eslint-plugin": "5.42.0",
@@ -48,7 +52,7 @@
     "eslint-plugin-autofix": "1.1.0",
     "eslint-plugin-import": "2.26.0",
     "eslint-plugin-jsx-a11y": "6.6.1",
-    "eslint-plugin-react": "7.31.10",
+    "eslint-plugin-react": "^7.31.10",
     "eslint-plugin-react-hooks": "4.6.0",
     "eslint-plugin-unused-imports": "2.0.0",
     "husky": "2.7.0",
