+++ src/Place/Models/PDB01MPlace/Contexts/PCT012MPlaceOne/ModelComponents/PMC01201GoogleMapPlaceOne/GoogleMapPlaceOne.tsx
@@ -1,34 +1,16 @@
-import { GoogleMap, Marker as MarkerWeb } from '@react-google-maps/api';
 import { useContext } from 'react';
-import { Platform } from 'react-native';
 import MapView, { Marker, PROVIDER_GOOGLE } from 'react-native-maps';
 
 import { PCT012MPlaceOne } from '@/Place/Models/PDB01MPlace/Contexts/PCT012MPlaceOne/MPlaceOne';
+import { styles } from './GoogleMapPlaceOneStyle';
 
 export const PMC01201GoogleMapPlaceOne = () => {
 	const { place } = useContext(PCT012MPlaceOne);
 
-	if (Platform.OS == 'web') {
-		const containerStyle = {
-			height: '300px',
-			width: '100%',
-		};
-
-		const centerPoint = {
-			lat: place.geometry.latitude,
-			lng: place.geometry.longitude,
-		};
-		const coordinate = { lat: place.geometry.latitude, lng: place.geometry.longitude };
-		return (
-			<GoogleMap mapContainerStyle={containerStyle} center={centerPoint} zoom={10}>
-				<MarkerWeb position={coordinate} title={place.name} key={place.place_id} />;
-			</GoogleMap>
-		);
-	}
 	const coordinate = { latitude: place.geometry.latitude, longitude: place.geometry.longitude };
 	return (
 		<MapView
-			style={{ width: '100%', height: 200 }}
+			style={styles.mapview}
 			provider={PROVIDER_GOOGLE}
 			initialRegion={{
 				latitude: coordinate.latitude,
