+++ src/Place/Models/PDB01MPlace/Contexts/PCT012MPlaceOne/ModelComponents/PMC01201GoogleMapPlaceOne/GoogleMapPlaceOneStyle.ts
@@ -1,0 +1,8 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+    mapview: { 
+        width: '100%',
+        height: 200 
+    },
+})
