+++ src/Place/Models/PDB01MPlace/Contexts/PCT012MPlaceOne/ModelComponents/PMC01201GoogleMapPlaceOne/GoogleMapPlaceOneStyle.web.ts
@@ -1,0 +1,6 @@
+export const styles = {
+    containerStyle: {
+        height: '300px',
+        width: '100%',
+    },
+}
