+++ src/Place/Pages/PPA002Place/Place.tsx
@@ -7,7 +7,7 @@ import { styles } from './PlaceStyle';
 
 import { BottomTabParamList } from '@/App';
 import i18n from '@/Common/Hooks/i18n-js';
-import { PMC01201GoogleMapPlaceOne } from '@/Place/Models/PDB01MPlace/Contexts/PCT012MPlaceOne/ModelComponents/PMC01201GoogleMapPlaceOne';
+import { PMC01201GoogleMapPlaceOne } from '@/Place/Models/PDB01MPlace/Contexts/PCT012MPlaceOne/ModelComponents/PMC01201GoogleMapPlaceOne/GoogleMapPlaceOne';
 import { PMC01203PlaceImage } from '@/Place/Models/PDB01MPlace/Contexts/PCT012MPlaceOne/ModelComponents/PMC01203PlaceImage';
 import { PCT012MPlaceOne } from '@/Place/Models/PDB01MPlace/Contexts/PCT012MPlaceOne/MPlaceOne';
 
@@ -55,7 +55,7 @@ export const PPA002Place = ({ route, navigation }: NativeStackScreenProps<Bottom
 					</View>
 				)}
 				numColumns={2}
-				// keyExtractor={(place) => place.place_id}
+			// keyExtractor={(place) => place.place_id}
 			/>
 		</View>
 	);
