+++ src/Place/Pages/PlacePageNavigator/PlacePageNavigator.web.tsx
@@ -9,7 +9,7 @@ import { PPA002Place } from '../PPA002Place/Place';
 
 import { BottomTabParamList } from '@/App';
 import { ENV } from '@/ENV';
-import { PCT011MPlacesListProvider } from '@/Place/Models/PDB01MPlace/Contexts/PCT011MPlacesList';
+// import { PCT011MPlacesListProvider } from '@/Place/Models/PDB01MPlace/Contexts/PCT011MPlacesList';
 
 export type PlaceStackParamList = {
 	PPA001Places: PlacesPropsInterface; // 本来は PagePropsInterface を設定する
@@ -20,7 +20,7 @@ const Stack = createNativeStackNavigator<PlaceStackParamList>();
 
 export const PlacePageNavigator = ({ navigation }: NativeStackScreenProps<BottomTabParamList, 'Place'>) => (
 	<PCT012MPlaceOneProvider parentDocRef={undefined} initialPlaceId="" language="ja">
-		<PCT011MPlacesListProvider>
+		{/* <PCT011MPlacesListProvider> */}
 			<LoadScript googleMapsApiKey={ENV.GCP_API_KEY}>
 				<Stack.Navigator initialRouteName="PPA001Places">
 					<Stack.Screen
@@ -31,6 +31,6 @@ export const PlacePageNavigator = ({ navigation }: NativeStackScreenProps<Bottom
 					<Stack.Screen name="PPA002Place" component={PPA002Place} initialParams={{ place_id: '', language: '' }} />
 				</Stack.Navigator>
 			</LoadScript>
-		</PCT011MPlacesListProvider>
+		{/* </PCT011MPlacesListProvider> */}
 	</PCT012MPlaceOneProvider>
 );
